requirements.txt
Django>=5.1
channels>=4.0
daphne>=4.0
ptyprocess>=0.7
requests>=2.31
python-dotenv>=1.0
django-cors-headers>=4.0
